+++ crates/fuel-core/src/graphql_api/required_fuel_block_height_extension.rs
@@ -1,6 +1,5 @@
-use super::database::ReadView;
+use super::block_height_subscription;
 
-use crate::fuel_core_graphql_api::database::ReadDatabase;
 use async_graphql::{
     extensions::{
         Extension,
@@ -23,11 +22,12 @@ use std::sync::{
     OnceLock,
 };
 
+use tokio::time::Duration;
+
 const REQUIRED_FUEL_BLOCK_HEIGHT: &str = "required_fuel_block_height";
 const CURRENT_FUEL_BLOCK_HEIGHT: &str = "current_fuel_block_height";
 const FUEL_BLOCK_HEIGHT_PRECONDITION_FAILED: &str =
     "fuel_block_height_precondition_failed";
-
 /// The extension that implements the logic for checking whether
 /// the precondition that REQUIRED_FUEL_BLOCK_HEADER must
 /// be higher than the current block height is met.
@@ -35,29 +35,83 @@ const FUEL_BLOCK_HEIGHT_PRECONDITION_FAILED: &str =
 /// the request data by the graphql handler as a value of type
 /// `RequiredHeight`.
 #[derive(Debug, derive_more::Display, derive_more::From)]
-pub(crate) struct RequiredFuelBlockHeightExtension;
+#[display(
+    fmt = "RequiredFuelBlockHeightExtension(tolerance: {})",
+    tolerance_threshold
+)]
+pub(crate) struct RequiredFuelBlockHeightExtension {
+    tolerance_threshold: u32,
+    timeout: Duration,
+    block_height_subscriber: block_height_subscription::Subscriber,
+}
+
+enum BlockHeightComparison {
+    TooFarBehind,
+    WithinTolerance,
+    Ahead,
+}
+
+impl BlockHeightComparison {
+    fn from_block_heights(
+        required_block_height: &BlockHeight,
+        current_block_height: &BlockHeight,
+        tolerance_threshold: u32,
+    ) -> Self {
+        if current_block_height.saturating_add(tolerance_threshold)
+            < **required_block_height
+        {
+            BlockHeightComparison::TooFarBehind
+        } else if current_block_height < required_block_height {
+            BlockHeightComparison::WithinTolerance
+        } else {
+            BlockHeightComparison::Ahead
+        }
+    }
+}
 
 impl RequiredFuelBlockHeightExtension {
-    pub fn new() -> Self {
-        Self
+    pub fn new(
+        tolerance_threshold: u32,
+        timeout: Duration,
+        block_height_subscriber: block_height_subscription::Subscriber,
+    ) -> Self {
+        Self {
+            tolerance_threshold,
+            timeout,
+            block_height_subscriber,
+        }
     }
 }
 
 pub(crate) struct RequiredFuelBlockHeightInner {
     required_height: OnceLock<BlockHeight>,
+    tolerance_threshold: u32,
+    timeout: Duration,
+    block_height_subscriber: block_height_subscription::Subscriber,
 }
 
 impl RequiredFuelBlockHeightInner {
-    pub fn new() -> Self {
+    pub fn new(
+        tolerance_threshold: u32,
+        timeout: Duration,
+        block_height_subscriber: &block_height_subscription::Subscriber,
+    ) -> Self {
         Self {
             required_height: OnceLock::new(),
+            tolerance_threshold,
+            timeout,
+            block_height_subscriber: block_height_subscriber.clone(),
         }
     }
 }
 
 impl ExtensionFactory for RequiredFuelBlockHeightExtension {
     fn create(&self) -> Arc<dyn Extension> {
-        Arc::new(RequiredFuelBlockHeightInner::new())
+        Arc::new(RequiredFuelBlockHeightInner::new(
+            self.tolerance_threshold,
+            self.timeout,
+            &self.block_height_subscriber,
+        ))
     }
 }
 
@@ -96,72 +150,130 @@ impl Extension for RequiredFuelBlockHeightInner {
         operation_name: Option<&str>,
         next: NextExecute<'_>,
     ) -> Response {
-        let view: &ReadView = ctx.data_unchecked();
-
-        let current_block_height = view.latest_block_height();
-
         if let Some(required_block_height) = self.required_height.get() {
-            if let Ok(current_block_height) = current_block_height {
-                if *required_block_height > current_block_height {
+            let current_block_height =
+                self.block_height_subscriber.latest_seen_block_height();
+
+            match BlockHeightComparison::from_block_heights(
+                required_block_height,
+                &current_block_height,
+                self.tolerance_threshold,
+            ) {
+                BlockHeightComparison::TooFarBehind => {
                     let (line, column) = (line!(), column!());
-                    let mut response = Response::from_errors(vec![ServerError::new(
-                        format!(
-                            "The required fuel block height is higher than the current block height. \
-                            Required: {}, Current: {}",
-                            // required_block_height: &BlockHeight, dereference twice to get the 
-                            // corresponding value as u32. This is necessary because the Display 
-                            // implementation for BlockHeight displays values in hexadecimal format.
-                            **required_block_height,
-                            // current_fuel_block_height: BlockHeight, dereference once to get the 
-                            // corresponding value as u32.
-                            *current_block_height
-                        ),
-                        Some(Pos {
-                            line: line as usize,
-                            column: column as usize,
-                        }),
-                    )]);
-
-                    response.extensions.insert(
-                        CURRENT_FUEL_BLOCK_HEIGHT.to_string(),
-                        Value::Number((*current_block_height).into()),
+                    return error_response(
+                        required_block_height,
+                        &current_block_height,
+                        line,
+                        column,
                     );
-                    response.extensions.insert(
-                        FUEL_BLOCK_HEIGHT_PRECONDITION_FAILED.to_string(),
-                        Value::Boolean(true),
-                    );
-
-                    return response
                 }
+                BlockHeightComparison::WithinTolerance => {
+                    let timeout = self.timeout;
+                    match await_block_height(
+                        &self.block_height_subscriber,
+                        required_block_height,
+                        &timeout,
+                    )
+                    .await
+                    {
+                        Ok(()) => {}
+                        Err(e) => {
+                            let (line, column) = (line!(), column!());
+                            tracing::error!(
+                                "Failed to wait for the required fuel block height: {}",
+                                e
+                            );
+                            return Response::from_errors(vec![ServerError::new(
+                                "Failed to wait for the required fuel block height",
+                                Some(Pos {
+                                    line: line as usize,
+                                    column: column as usize,
+                                }),
+                            )]);
+                        }
+                    }
+                }
+                BlockHeightComparison::Ahead => {}
             }
-        }
+        };
 
         let mut response = next.run(ctx, operation_name).await;
 
-        // TODO: After https://github.com/FuelLabs/fuel-core/pull/2682
-        //  request the latest block height from the `ReadDatabase` directly.
-        let database: &ReadDatabase = ctx.data_unchecked();
-        let view = database.view();
-        let current_block_height = view.and_then(|view| view.latest_block_height());
-
-        if let Ok(current_block_height) = current_block_height {
-            let current_block_height: u32 = *current_block_height;
+        let current_block_height =
+            self.block_height_subscriber.latest_seen_block_height();
+        // Dereference to display the value in decimal base.
+        let current_block_height: u32 = *current_block_height;
+        response.extensions.insert(
+            CURRENT_FUEL_BLOCK_HEIGHT.to_string(),
+            Value::Number(current_block_height.into()),
+        );
+        if self.required_height.get().is_some() {
             response.extensions.insert(
-                CURRENT_FUEL_BLOCK_HEIGHT.to_string(),
-                Value::Number(current_block_height.into()),
+                FUEL_BLOCK_HEIGHT_PRECONDITION_FAILED.to_string(),
+                Value::Boolean(false),
             );
-            // If the request contained a required fuel block height, add a field signalling that
-            // the precondition was met.
-            if self.required_height.get().is_some() {
-                response.extensions.insert(
-                    FUEL_BLOCK_HEIGHT_PRECONDITION_FAILED.to_string(),
-                    Value::Boolean(false),
-                );
-            }
-        } else {
-            tracing::error!("Failed to get the current block height");
         }
 
         response
+    }
+}
+
+fn error_response(
+    required_block_height: &BlockHeight,
+    current_block_height: &BlockHeight,
+    line: u32,
+    column: u32,
+) -> Response {
+    let mut response = Response::from_errors(vec![ServerError::new(
+        format!(
+            "The required fuel block height is higher than the current block height. \
+            Required: {}, Current: {}",
+            // required_block_height: &BlockHeight, dereference twice to get the
+            // corresponding value as u32. This is necessary because the Display
+            // implementation for BlockHeight displays values in hexadecimal format.
+            **required_block_height,
+            // current_fuel_block_height: &BlockHeight, dereference twice to get the
+            // corresponding value as u32.
+            **current_block_height
+        ),
+        Some(Pos {
+            line: line as usize,
+            column: column as usize,
+        }),
+    )]);
+
+    response.extensions.insert(
+        CURRENT_FUEL_BLOCK_HEIGHT.to_string(),
+        Value::Number((**current_block_height).into()),
+    );
+    response.extensions.insert(
+        FUEL_BLOCK_HEIGHT_PRECONDITION_FAILED.to_string(),
+        Value::Boolean(true),
+    );
+
+    response
+}
+
+async fn await_block_height(
+    block_height_subscriber: &block_height_subscription::Subscriber,
+    block_height: &BlockHeight,
+    timeout: &Duration,
+) -> anyhow::Result<()> {
+    tokio::select! {
+        biased;
+        block_height_res = block_height_subscriber.wait_for_block_height(*block_height) => {
+            block_height_res.map_err(|e| {
+                anyhow::anyhow!(
+                    "Failed to wait for the required fuel block height: {}",
+                    e
+                )})
+        },
+        _ = tokio::time::sleep(*timeout) => {
+            Err(anyhow::anyhow!(
+                "Timeout while waiting for the required fuel block height: {}",
+                block_height
+            ))
+        }
     }
 }